# Conflicts:
#	common-interfaces/src/main/java/org/eea/interfaces/vo/dataflow/DataFlowVO.java
#	dataflow-service/src/main/java/org/eea/dataflow/persistence/domain/Dataflow.java
#	dataflow-service/src/main/java/org/eea/dataflow/service/impl/DataflowServiceImpl.java
#	dataset-service/src/main/java/org/eea/dataset/service/BigDataDatasetService.java
#	dataset-service/src/main/java/org/eea/dataset/service/DatasetService.java
#	dataset-service/src/main/java/org/eea/dataset/service/impl/BigDataDatasetServiceImpl.java
#	dataset-service/src/main/java/org/eea/dataset/service/impl/DatasetServiceImpl.java
#	dataset-service/src/test/java/org/eea/dataset/io/kafka/commands/CheckBlockersDataSnapshotCommandTest.java
#	frontend-service/src/conf/messages.en.json
#	frontend-service/src/entities/Dataflow.js
#	frontend-service/src/repositories/CitizenScienceDataflowRepository.js
#	frontend-service/src/repositories/DataflowRepository.js
#	frontend-service/src/services/CitizenScienceDataflowService.js
#	frontend-service/src/services/DataflowService.js
#	frontend-service/src/services/_utils/DataflowUtils.js
#	frontend-service/src/views/Dataflow/Dataflow.jsx
#	frontend-service/src/views/Dataflow/_components/BigButtonList/_functions/Hooks/useBigButtonList.js
#	frontend-service/src/views/Dataflow/_components/BigButtonListRepresentative/_functions/Hooks/useBigButtonListRepresentative.js
#	frontend-service/src/views/_components/ManageDataflow/ManageDataflow.jsx
#	frontend-service/src/views/_components/ManageDataflow/_components/ManageDataflowForm/ManageDataflowForm.jsx
#	frontend-service/src/views/_components/ManageDataflow/_functions/Reducers/reportingDataflowReducer.js
#	frontend-service/src/views/_components/ManageReferenceDataflow/ManageReferenceDataflow.jsx
#	validation-service/src/main/java/org/eea/validation/util/ValidationHelper.java

+++ dataset-service/src/main/java/org/eea/dataset/controller/DatasetSchemaControllerImpl.java
@@ -127,8 +127,6 @@ public class DatasetSchemaControllerImpl implements DatasetSchemaController {
   @Lazy
   @Autowired
   private BigDataDatasetService bigDataDatasetService;
-    @Autowired
-    private DataFlowControllerZuul dataFlowControllerZuul;
 
     /**
    * Creates the empty dataset schema.
@@ -584,10 +582,19 @@ public class DatasetSchemaControllerImpl implements DatasetSchemaController {
       if(Boolean.TRUE.equals(isBigDataFlow)){
         updateMaterializedViews = false;
       }
-      if (BooleanUtils.isTrue(isBigDataFlow) && !StringUtil.isNullOrEmpty(tableSchemaVO.getIdTableSchema())) {
+      //Deleting data from the old table and creating a new table should only occur when renaming a table, meaning when
+      //getNameTableSchema is not blank.
+      if (BooleanUtils.isTrue(isBigDataFlow)
+              && !StringUtil.isNullOrEmpty(tableSchemaVO.getIdTableSchema())
+              && StringUtils.isNotBlank(tableSchemaVO.getNameTableSchema())) {
         bigDataDatasetService.deleteTableData(datasetId, dataflowId, null, null, tableSchemaVO.getIdTableSchema(), null, false);
-      }
+       }
       dataschemaService.updateTableSchema(datasetId, tableSchemaVO, updateMaterializedViews);
+
+      //Creating an empty table for the new table name is required for ValidateQCs functionality to work correctly.
+      if (StringUtils.isNoneBlank(tableSchemaVO.getNameTableSchema())) {
+        bigDataDatasetService.createEmptyTablesForSpecificTableSchema(datasetId, tableSchemaVO.getIdTableSchema());
+      }
     } catch (EEAException e) {
       LOG.error("Error updating table schema for datasetId {}. Message: {}", datasetId, e.getMessage(), e);
       if (e.getMessage() != null
@@ -832,6 +839,8 @@ public class DatasetSchemaControllerImpl implements DatasetSchemaController {
           @ApiParam(value = "Field schema object") @RequestBody FieldSchemaVO fieldSchemaVO) {
 
 
+    final boolean isBigData = dataflowControllerZuul.isBigDataflowDataset(datasetId);
+
     if (null != fieldSchemaVO.getName()) {
       if (fieldSchemaVO.getName().chars().anyMatch(Character::isWhitespace)) {
         throw new ResponseStatusException(HttpStatus.BAD_REQUEST, EEAErrorMessage.FIELD_NAME_WHITESPACES);
@@ -859,7 +868,7 @@ public class DatasetSchemaControllerImpl implements DatasetSchemaController {
         dataschemaService.updateForeignRelation(datasetId, fieldSchemaVO, datasetSchema);
 
         // Clear the attachments if necessary
-        if (Boolean.TRUE.equals(
+        if (!isBigData && Boolean.TRUE.equals(
                 dataschemaService.checkClearAttachments(datasetId, datasetSchema, fieldSchemaVO))) {
           datasetService.deleteAttachmentByFieldSchemaId(datasetId, fieldSchemaVO.getId());
         }
